+++ src/datasentry/__init__.py
@@ -2,6 +2,6 @@
 
 from datasentry.client import DataSentry
 
-__version__ = "0.24.0"
+__version__ = "0.25.0"
 
 __all__ = ["DataSentry", "__version__"]